inline method refactoring: fix missing '$' in variable name
Summary:
Before this change, when inlining a method with a `return` statement we assign the return value to a variable `res` or `res_`, or `res__`, etc (bug T164085986)

After this change, when inlining a method with a `return` statement we assign the return value to a variable `$res` or `$res_`, or `$res__`, etc.

Reason for this change: variables in Hack start with `$`.

Reviewed By: patriciamckenzie

Differential Revision: D49365854

fbshipit-source-id: 13875a19830dd2ce197f7f16e683d8fc6f98ab28