Add support for using as experimental feature
Summary:
Exhaustivity checking in Hack misses many cases (see [this post](https://fb.workplace.com/groups/hackforhiphop/permalink/24988113060810640/) or D47890092). This diff continues the implementation of a new strict switch gated behind an option.

It does so by changing how the feature is gated/enabled. Previously, it was gated behind a typechecker config option. This diff adds the ability to enable it via `<<file: __EnableUnstableFeatures('strict_switch')>>` and `<<__StrictSwitch>>` annotation on functions and methods [as discussed here](https://fb.workplace.com/groups/hacklang.design/permalink/6542879245803959/).

This diff is complicated slightly by the fact there is no way to access the annotations of the enclosing function from a (switch) statement and the environment the visitor provides. Hence, it overrides the function and method visitors to check for that, and uses a mutable variable to pass that information to the statement visitor.

Non-goals. This diff does not switch over all the tests to gate via the unstable features annotation; nor does it remove the typechecker option for the gating. Both of these are done in the subsequent diff.

Reviewed By: dlreeves

Differential Revision: D49376438

fbshipit-source-id: 920bd7c8deac74932b99dc7c836356d0f5334d12