More efficient Keyset\diff implementation
Summary:
This new implementation of Keyset\diff has some benefits. The new implementation was tested internally at Meta and showed CPU and latency improvements. We needed to test/prove it is an improvement, because we weren't sure of it mainly due to not being certain of the realistic efficiency of `unset`.

This has few benefits:
* Less space complexity, because we don't put `$second` and `$rest` into a new keyset via `!$rest ? keyset($second) : union($second, ...$rest);`.
* Potentially return early if `$first` becomes empty due to removals.
* This is also potentially more efficient because if `$first` is already a `keyset`, then there will be no iterations over `$first` (if it's not a `keyset`, then converting it to a `keyset` needs to iterate it to do so. The original implementation would iterate `$first` via the `filter` function.
`$second` and `$rest` are iterated in the old implementation and new implementation (except when there's an early return).

Reviewed By: admdikramr

Differential Revision: D49383300

fbshipit-source-id: 197444cc9770e7e3843376d441efdc978b37b885