Introduce ClassGraph abstraction in HHBBC
Summary:
Right now in HHBBC when you want information related to the class
hierarchy, you have two options: You can iterate over a class'
subclass list and calculate it yourself, or you can pre-calculate it
during the build-subclass-lists phase. This bloats ClassInfo because
that information has to be stored even if it's never needed (and some
of this information can be quite large). Calculating it on-demand from
the subclass list can be expensive (because the subclass list can be
very large). The situation is even worse when running in an
extern-worker job, because the subclass list only contains names and
there's no guarantee that the actual class will be present in that
job.

Indeed, the mere existence of the subclass list is problematic. As
said above, the lists can become very large, especially for wildy used
interfaces. Moreover, they are memory wasteful. Each class stores the
list of all of it's transitive children. This adds up in deep
hierarchies.

ClassGraph is an attempt to improve some of these issues. An instance
of ClassGraph can be thought of as a "pointer" into a global graph
representing the class hierarchy of the program under analysis. The
graph representation saves memory because the transitive children or
parents aren't flattened, avoiding duplication. All of the information
currently inferred from the class hierarchy can be inferred using
ClassGraph by simple walks over the edges. The N different specialized
sets of lists that need to be pre-calculated are no longer necessary.

The real savings come from when running in extern-worker jobs. If the
entire program-wide hierarchy had to be serialized, it would be
prohibitively expensive. However, when a ClassGraph is serialized,
only the subset of the hierarchy which is relevant to that ClassGraph
is serialized. Right now that's defined as all transitive parents of
the class, all transitive children of the class, and all transitive
parents of the children. This leaves out the children of any
transitive parents, which keeps the graph from exploding. Moreover,
when serializing or deserializing, we only ever serialize a node in
the graph once, so sharding is maintained.

An advantage to this approach is that ClassGraph is self
contained. Right now within an extern-worker job, in order to perform
operations on the class hierarchy, one does not need a subclass list,
but all ClassInfos on that subclass list. The subclass list itself
does not provide enough information. This is awkward. In comparison,
the ClassGraph encodes all of the relevant information
directly. Having an instance of a ClassGraph in a job is sufficient to
perform most (all?) queries on the class hierarchy. If we embded
ClassGraph within a Type (not yet done), this means Type is
self-contained, which *majorily* simplies converting the analysis
phase to extern-worker.

Move ClassInfo2 over to use ClassGraph directly. ClassInfo still uses
an explicit subclass list, but it's generated from the ClassInfo2
ClassGraph. We'll use ClassGraph directly in follow up diffs.

Reviewed By: jano

Differential Revision: D49359322

fbshipit-source-id: 291ff67654670223c9baf41437e70c01e0db2328