Ensure request is destroyed in EventBase thread on !getShouldStartProcessing()
Summary:
Before this diff, an inflight `ResponseChannelRequest` will be destroyed on the Python `asyncio` executor thread, not on the correct `EventBase` thread, if `getShouldStartProcessing()` returns false. This is a problem since `ResponseChannelRequests` are not thread-safe. We observed use-after-frees in Inference Platform's Python predictor when the request was destroyed on the `asyncio` executor thread, and the Thrift queue timeout callback fired shortly after on the request's `EventBase` thread.

Since the `HandlerCallback` destructor ensures that the request is destroyed on the request's `EventBase` thread, this diff simply moves the request into a `HandlerCallback` object before returning early in case `getShouldStartProcessing()` return false.

Reviewed By: Filip-F

Differential Revision: D49351539

fbshipit-source-id: 4654cbacb24a7a827686ac97ab85d3037fe11b6e