non-optional field should not be ensured to non-optional default
Summary:
Static patch will ignore ensure operation since those fields always have value.

However, ensuring non-optional field is still useful for dynamic patch, since in dynamic patch, everything is optional, we can remove a field, or, for terse field, they might not be serialized. In this case we need to ensure them first so that we can patch them later (we won't patch fields that are not set).

When we remove a non-optional field, or when non-optional field doesn't exist in the serialized data, it will have the default value, thus we must always ensure non-optional field to default value.

You might wonder, if we ignore user provided ensure value, won't this be misleading? That's already the case today for static patches when we ignore EnsureOp at all. Ideally, we should remove such a use case, though it would require some codemod which we can do later.

Reviewed By: thedavekwon

Differential Revision: D49354694

fbshipit-source-id: 5f3e96fba811609a4c13263af27db5e69167d9e3