Make ListPatch::prepend private
Summary:
This is not used except for internal APIs (e.g., `diff`).

After this diff, when we read patch that contains `ListPatch::prepend`, it still works for backward compatibility, but people can't create patch with `ListPatch::prepend`.

Once we confirmed no one is creating `ListPatch::prepend` anymore, we will remove this Op.

The motivation is to keep ListPatch as small as possible, so that it's easier to evolve in the future.

Reviewed By: thedavekwon

Differential Revision: D49348997

fbshipit-source-id: 82246e8feea70452ab16e23f8c4fed08a0af75f3