Fix Behavior of Case Types Under Implicit Pess
Summary:
# Context
D47351127 intended to prevent case types from being used as the base type of an enum or enum class due to limitations in the runtime. Though the check was tested, it was not tested when the type checker runs under implicit pessmisation. In that case the base type could be wrapped in a like type, leading to not reporting an error.

# Mitigation
I moved the check to be reported as part of the traversal we do in `is_valid_base`. Other option is to move this to a tast check, but it feels simpler to do it at this point.

Reviewed By: vassilmladenov

Differential Revision: D49428186

fbshipit-source-id: 9fb0a3ae57b844ade74d71ddd28e8c66b6d0ba16