Thread through print.rs
Summary: Same as D49296892; we pass down the scope from the emitter to the expression printer instead of making a custom env for it. Was able to clean up some code in the process e.g. the `repr(C)` for HhasBodyEnv appears to be no longer necessary, and the ExprEnv can be a simple type alias instead of a struct.

Reviewed By: aorenste

Differential Revision: D49343984

fbshipit-source-id: d9c1c4e960b25c5a356fe5f354697574f1be4c83