Fix unsafe_to_raw_pos in typing attribute checks
Summary:
When checking attributes, we need to check that each attribute has a defined attribute class, and that the attribute class correctly implements the corresponding marker interface. Therefore, each attribute is essentially a classname access.

However, instead of creating a proper classname access, we pass in a raw decl position to new_object instead of the position where the attribute class is used(i.e. the attribute position), which can result in errors outside of a file. Instead, we should generate a correct CID using the attribute position.

Reviewed By: ljw1004

Differential Revision: D49541443

fbshipit-source-id: 278ad01e3bb4f520223cd6102a80a2419a251a74