Remove references to RPCServer
Summary:
Just some housekeeping. Remove the XboxRequestHandler layer on top of
RPCRequestHandler, then rename RPCRequestHandler to XboxRequestHandler. Update
references.

Previously this landed as D46748801 and caused S351778. I've landed a new unit test against the bug from the original diff in D47859374 (it passes on this diff) as well as run some of the repros from the SEV to validate that the issue isn't there anymore.

Reviewed By: ricklavoie

Differential Revision: D47880415

fbshipit-source-id: 813e160ad6f43267f3c17c81f60a82ac2e56e10b