log activity_id, not tracking_id
Summary:
In Hh_server_events we log a field `tracking_id`, a unique ID for this particular incoming LSP message (be it request or notification). For requests, we form it by concatenating the LSP id.

We've never used this field EVER.

In this diff, I changed it to store **activity_id** in that column. The activity_id is whatever was present (if any) in the LSP request of this form:
```
[2023-09-22 16:34:20.232] [Lsp#saAUYEUfAL] JSON={
  "jsonrpc":"2.0",
  "method":"textDocument/didOpen",
  "params":{
    "textDocument":{
      "uri":"file:///data/users/ljw/www/flib/intern/entity/dog_park/EntDog.php",
      "languageId":"hack",
      "version":1,
      "text":"<?hh\n// ...\n"
    }
  },
  "activityKey":{
    "name":"ActivateExtension",
    "id":"8e37be1c-c2e5-4a0e-9a89-aeedf35e661a",
    "$akey$":true
  }
}
```

Not all LSP requests include an activityKey.id, in which case the `lsp_tracking_id` column will simply be the string "_".

Note: We still use `tracking_id` in various places in the code, e.g. Hh_logger.log will write that tracking-id to the clientLsp log and the clientIdeDaemon log. I still retain `tracking_id` for that purpose, although now I've lengthened it to include activity_id. But we no longer log it to scuba.

Throughout, I'm concerned about the volume of data we write to scuba. That's why I wanted to keep our scuba logging minimal here. This diff will end up storing slightly longer strings for many LSP methods (because VSCode tends to send activityKey.id which are ~32bytes while our tracking-ids were closer to ~15bytes), and storing shorter strings for the LSP methods like didChange which don't have an activityKey.id. So I think it will be more or less a wash.

Reviewed By: patriciamckenzie

Differential Revision: D49565977

fbshipit-source-id: 18cada8f872932558f2159b49c46714ea8f92d93