Add support for arraykey class constants
Summary:
Exhaustivity checking in Hack misses many cases (see [this post](https://fb.workplace.com/groups/hackforhiphop/permalink/24988113060810640/) or D47890092). This diff continues the implementation of a new strict switch gated behind an option.

It does so by adding the a `String` variant to the `Value.t` datatype in `Strict_switch_check.ml` and `Typing_error.mli`. First, the Hack `string` type gets mapped to `String` and `arraykey` is to be mapped to `Value.[Int; String; Other]`. Next, string literals and class consts are bucketed into the new `Value.String` partition if present. Finally, strings literals are checked for redundancy.

Non-goals. This diff does not intend to add support for enums. Support for this will be added in later diffs.

Reviewed By: dlreeves

Differential Revision: D49539119

fbshipit-source-id: 86282aa93266b1109c91d4c331db8ce63c6dcb6d