Factor out enum helper function
Summary:
Exhaustivity checking in Hack misses many cases (see [this post](https://fb.workplace.com/groups/hackforhiphop/permalink/24988113060810640/) or D47890092). This diff continues the implementation of a new strict switch gated behind an option.

It does so by factoring out the logic used to detect whether a newtype is actually an enum of some sort. This will enable the support for enums in subsequent diffs.

Reviewed By: dlreeves

Differential Revision: D49539806

fbshipit-source-id: a1d87dbfb1ba1b6b94dde6796f2120f72256a7bd