Force lazy values in Tast.saved_env before marshalling
Summary: hh_server sometimes returns TASTs to hh_client. For that it needs to marshal it, which does not work for functional values. Lazy values can be functional so we need to force them in the TAST before marshalling. We already do that, but we forgot to force lazy values in Tast.saved_env. This diff does that.

Reviewed By: madgen

Differential Revision: D49641111

fbshipit-source-id: 902f770e5902869a1a8cc5c13fce4eb0ebb7f684