delete NamespaceSearchService
Summary:
The NamespaceSearchService was used to reconstruct a hierarchical namespace view out of the namespaces registered in the www.autocomplete database.

(Note: we never provided autocomplete for namespaces defined in the local delta, as far as I can see?)

Anyway, we no longer use www.autocomplete database. We instead use www.hack.light, and a different schema, one which offers namespaces directly. Therefore, NamespaceSearchService is defunct.

By the way, NamespaceSearchService did one other thing, which was to make sure the namespace-aliases from .hhconfig are represented in autocomplete suggestions. But this is a pretty small uses, and doesn't need the full weight of NamespaceSearchService, so I implemented it more simply in SymbolIndex.ml

Reviewed By: nt591

Differential Revision: D49478296

fbshipit-source-id: f534cafbbc23984ab9103c59ec6bc3e75df13837