Gracefully Handle Cyclic Types in Typing_case_types
Summary:
# Context
Typing_case_types will walk the type structure recursively in an attempt to reduce it to a set of possible run time tags that are associated with the type. The current implementation is vulnerable to infinite looping and stack overflow segfaults if it encounters a cycle when traversing a type structure. This can occur if there is a cycle in:
* Generic bounds (i.e. `T1 as T2, T2 as T1`)
* Bounds on a newtype
* Sealed List or Requirements
* Expanding a case type

This diff hardens the implementation in Typing_case_types to make conservative approximations when encountering cycles instead of producing divergent behavior.

# Mitigation
Most of the tracking was already in place when D48653810 started computing reasons for why a type was mapped to a certain set of runtime tags. In particular we maintain a trail each time we recursively call into the `DataType.fromTy` function. I changed `DataTypeReason.t` to include a `Typing_defs.Type_expansion.t` to keep track of expansions when they occur. When the trail is updated it uses `Type_expansion` to check for cycles and reports an error a position is returned from `Type_expansion`. For now this should never occur because we do not set `report_cycle` when creating the `Type_expansion.t` in `DataType.fromHint`/`fromTy`. This is because we still do cycle detection as part of the general well-formedness checks inside typing_typedef.

Since updates to the trail can fail due to encountering a cycle, it will now return `Result.t` where `Result.Error` returns information about the cycle. The usage of `Result.t` forces any recursive calls that update the trail to properly handle the possibility of encountering a cycle. This is standardized in `DataType.cycle_handler` which requires supplying a default `DataType.t` for when a cycle is encountered. In most cases this will be the set of all possible runtime tags (`mixed`), but will be the set of all object tags within `DataType.Class.to_datatype` or `InstanceOf I` for some interface `I` if the cycle was encountered while recursing on the sealed or required list.

# Related Issue
As part of testing this change I discovered another potential infinite loop that can occur when localizing a case type. This will only occur if the case type contains a variant that is a case type defined with a cycle (see hphp/hack/test/typecheck/case_type/case_type_cyclic_case_type.php for an example of this). This will occur because there is a cyclic dependency between localizing and subtyping.

* When a case type is opened, the type it points to will contain a `Tunion`.
* When we localize a union type we call into `Typing_union.union_list` to normalize the union (i.e. change `int | float` to `num`)
* Typing_union will call `Typing_utils.is_sub_type_for_union_i` as part of normalizing and simplifying a union
* When a case type is the super type in `Typing_subtype.simplify_subtypei` we fallback to opening the case type which requires localizing the type

We do not thread through the `Typing_defs.expand_env` through all of these layers. This affects case types because we don't attempt to fully expand case types during localization, unlike type aliases and new types. These two things combine to cause the infinite looping when a cycle case type is encountered. The fix itself is to include the name of the case type we are expanding when opening the case type during subtyping. In that case I will say the subtyping proposition is invalid.

Reviewed By: vassilmladenov

Differential Revision: D49560162

fbshipit-source-id: 519aef28dc52b2f705d3625005c0fa4370acc34e