Deduplicate side exits from inlined frames
Summary:
The current implementation spills all inlines frames linearly whenever we need
to side exit. This results in a lot of code duplication in the cold and frozen
sections.

Let's instead make all side exits in an inlined frame flow into the shared side
exit block of the parent frame, so that side exits nicely compose into a tree
structure.

Since this reverses the order in which frames are spilled and the address of
the side exit is specific to each initial point of side exit, the mechanism
needed to be rearchitected. We now side exit one frame at a time.

When a side exit is requested, we force materialization of the frame's stack,
bind the address of the side exit and pass it to the shared side exit block of
the inlined frame. That block calls the inlinedSideExit stub that initialize
the ActRec and jumps to the bound address. Otherwise this acts like a regular
VM call, so when that function returns, we obtain the return value and push it
to the stack.

Since we call into each inlined frame to continue its execution, this fixes the
stack imbalance of return address branch prediction.

The initial implementation continues to side exit recursively. However, it
unblocks further improvements, such as rejoining the main execution code path,
or at least Awaitable suspend path. However, this needs to be done carefully to
avoid pessimization of the hot path.

There is an interdependence of side exits and catch blocks. InlineSideExit
instructions need to have a catch block similar to Call or CallFuncEntry. When
we can't optimize away these catch blocks, we have to emit code to spill frames
in the legacy way anyway. Some of that can be fixed to allow catch blocks to
flow back to Hack catch blocks, most likely triggering a side exit. Elimination
of usage of spillInlinedFrames() could reduce the code size substantially.

While this diff deduplicates ActRec spills (~8% code size reduction), it
doesn't do much for spills of locals. However, it provides opportunity to build
a pass that will sink these spills towards the tree root. This is a ~50% code
size reduction opportunity.

Reviewed By: ricklavoie

Differential Revision: D49572290

fbshipit-source-id: bb022d0462594ec1ed4d0d1bcf666f3657c04817