Do not store m_func and m_callOffAndFlags to inlined frames
Summary:
Nobody reads these. Except of CreateAFWH, which rely on existence of m_func.
So set it here. And set it when we spill inlined frames.

This is usually what was happening anyway, so let's save that unnecessary
reshuffling and make the code easier to argue about.

Allows further improvements of inlining side exit logic.

Reviewed By: ricklavoie

Differential Revision: D49576837

fbshipit-source-id: 1a1faf4f44b7379ac868380d6153163065b08342