Fix and simplify if{Then,}{Else,} + cond irgen
Summary:
Do not attempt to emit next and taken branches if they are unreachable.

kshaunak made this mostly benign a while ago by ignoring instructions appended
to unreachable blocks. But the emitting code still runs and works with false
data, such as stack depth and BC markers. And these can leak into codegen that
does not check for reachability, such as makeExit() and everything else using
the BlockPusher logic.

This was especially dangerous in case of ReqBindJmp logic, which might
initialize SrcRec for a given SrcKey and emit stubs with incorrect stack depth.
I do not have evidence that this was happening, but it could explain occasional
handleTranslate() crashes. However, my future diffs triggered this.

In combination with various bespoke logic, this also allowed stack depth to go
negative. Add some asserts to catch similar issues in the future.

Reviewed By: ricklavoie

Differential Revision: D16968361

fbshipit-source-id: 71712e19a5cd4d62dc073162b703e08cbd6df968