TAST hashing: ignore positions coming from decls
Summary:
If nothing else than these positions change, then there's no need to recheck the file (unless it has errors).

TAST hashes are currently used to determine if a file should have been rechecked upon a change: if the TAST hash changed after a repo change, then the file for that TAST must be included in the fanout. However, we don't care about changes of positions from decls. These only impact errors lines, but won't create new errors, and we always recheck files with errors.

Issues with this diff: for any type x, you'd expect hash_x to hash everything in an x. If not, there's no way to write a doc block for hash_x, to explain why we hash this way. Suggestions welcome.

Differential Revision: D49688471

fbshipit-source-id: 8ffd0b07b834f7c303f54a90f23c98043a60fd3e