fix Lwt.wakeup_later bug, for $/cancel handling
Summary:
The IDE was sending these messages:
```
textDocument/references (id 10)
...
$/cancelRequest (cancel id 10)
textDocument/references (id 11)
```
The handler for `$/cancelRequest` requested cancellation of the find-refs shellout by doing `Lwt.wakeup_later cancellation_token ()`. The cancellation had been earlier passed to Lwt_utils.exec_checked, where it's handled by sending a SIGKILL to the process, but that's all it does -- we still have to wait for the normal course of affairs, of the process receiving a SIGKILL, indicating that it has terminated, us receiving the termination. In other words, even though we received `$/cancelRequest`, the shellout is still ongoing.

But then a second `textDocument/references` request came in. The handler for this will see an ongoing shellout, and so will (1) trigger its cancellation token once again so that the underlying process can die, (2) send an LSP response for request id 10, (3) remove the old `current_shellout` structures and replace them a new one and its new cancellation token.

The problem is that I was using `Lwt_wakeup_later cancellation_token ()` to trigger cancellation token. But the behavior of `Lwt_wakeup_later` is to throw an exception if it had already previously been triggered (as in this case).

The consequence was that our second request got an exception, and couldn't complete, and the old shellout structures I think remained in place.

Reviewed By: mheiber

Differential Revision: D49734259

fbshipit-source-id: 775111891bf83647d568700eed90993ab53af049