fix bug in flushing remaining find-refs items
Summary:
The design of find-refs is that
1. we shell-out to `hh --ide-find-refs-by-symbol3 <stream_file>`
2. we `tail -f` that stream file, and for each ref that it reports, we send a $/progress LSP message to VSCode
3. when the shellout process has completed, we unlink the file, remove our record of the `partialResultToken` and `stream_file`, and that's the end of it.

There's a race inherent in this! What if the process writes the final two items to the stream then terminates -- will we stop our "tail -f" too early hence missing the final items? or will we catch them before we catch process termination? Max had observed the race in practice, where sometimes he'd get all the $/progress he expected but sometimes he wouldn't.

(unix pipes don't have this problem because EOF is what you get when you read the pipe and it's finished. unix `tail -f` *does* have this problem because it never knows when the file has finished; it relies upon you to Ctrl+C it).

I attempted to fix this in D48714783 by making the shellout-complete handler first read to the end of the file. After all, it's impossible for the stream-file to grow any longer after the shellout process has terminated.

However I didn't write any tests, and I didn't know how to because it's all about races. And it turns out my fix had an error. Specifically, the shellout-complete handler was looking at the *current state* of `partialResultToken` in order to know how to produce its $/progress notification. That was wrong, because step 3 had already zeroed out our record of "currently underway partialResultToken". Here's an example log which illustrates the problem: P825722869
```
next event: Shell_out_complete
shell-out completed. [{"name":"IWorkBot::TMessagePayload","filename":"/data/users/mheiber/www/flib/in"}]
Refs-file: unexpected items when stream isn't open
```
To explain: `next event` happened when we saw that the process had terminated, and at this point we zero out the LSP state machine's record of any current shellout, and we store information about the just-finished shellout in the event. There is no subsequent `next event`... the `Refs-file: unexpected items` error is coming from within `handle_shell_out_complete`, as it mops up the last remaining items in the stream.

This diff fixes the problem. It makes it so we send $/progress notifications based on the partialResultToken that was in effect at the moment of shellout.

Alas I still don't have any ideas how to write tests for this.

Reviewed By: mheiber

Differential Revision: D49753494

fbshipit-source-id: 216298344987cd6bdf7ad5e765f6235cb6dfded1