fix discrepancy when clear optional patch
Summary:
Today when clearing the patch in patch prior and ensuring it later, there is a discrepancy between static and dynamic patch, since clear is set to intrinsic default in dynamic patch, but remove the field in static patch.

```

  Opt foo;
  foo.field() = 10;
  auto dynFoo = protocol::asValueStruct<type::struct_t<Opt>>(foo);

  OptPatch patch;
  patch.patchIfSet<ident::field>().clear();
  patch.ensure<ident::field>(1);

  // Apply patch statically
  patch.apply(foo);
  EXPECT_EQ(foo.field(), 1);

  // Apply patch dynamically
  protocol::applyPatch(patch.toObject(), dynFoo);
  EXPECT_EQ(dynFoo.as_object()[FieldId{1}].as_i32(), 1);
```

This diff fixed the issue by using assign in patch after instead of ensuring it (so that patch after will always override the value even dynamic patch did not remove the field).

Reviewed By: thedavekwon

Differential Revision: D49354695

fbshipit-source-id: ee8ac5c5c821403af7785127ba7dd04d64fd76d3