Fix pathological handling of constants wrt phis
Summary:
RefineTmps may see a CheckType InitNull and refine the original value into a
non-DefConst SSATmp holding a constant. Then when the branches rejoin, it will
use that SSATmp in a phi, which will always extend the phi, even in presence of
an equivalent set of inputs (except of using a DefConst). After refineTmps
pass, other passes will clean this up, but sometimes in a way that makes
more work for refineTmps pass downstream. If an equivalent pattern exists
downstream serially N times, the whole set of passes rerun N times until
convergence.

Let's fix this by converting a const SSATmp into an actual DefConst.

Reviewed By: ricklavoie

Differential Revision: D49804201

fbshipit-source-id: 68735053aa95ea2e1d2bc9d0b7b58377644ad53b