Remove redundant check in method-lookup.cpp
Summary:
We don't need to do deployment checks when looking up methods: in order to lookup a method we need to have resolved a class already, so we should be able to instead do the check on the class.

Returning null here on a method lookup when the class lookup is earlier actually is worse than just a code smell; it returns null and causes irgen to go down the slow path in repo mode when calling static methods on classes, which results in us being unable to optimize module and deployment boundary checks down the line.

This diff deletes all the packageInfos from method-lookup.cpp and its ancestors. We add a few extra checks to ResolveClsMethodD and its reified counterpart, as those were depending on the method lookup to raise warnings for deployments.

Reviewed By: oulgen

Differential Revision: D49740396

fbshipit-source-id: 14eba5c5f7c053eef95b8da46dcfa6ca7eea9087