Merge various pieces of frame-specific inlining state together
Summary:
InlineState maintained 3 different stacks of frame-specific data. Merge them
into the InlineFrame struct. Build it at the beginning of beginInlining()
instead of a combination of translate-region.cpp and end of beginInlining().
This deduplicates previous InlineReturnTarget construction logic.

Preserve the difference in handling profiling counts and add a FIXME that might
pique someone's curiosity to either improve it for a win, or kill it.

Reviewed By: ricklavoie

Differential Revision: D49718770

fbshipit-source-id: f7ff717db9e48a9b707a958e4518fdf9c9a95d5d