Refactor enum helper logic
Summary:
Exhaustivity checking in Hack misses many cases (see [this post](https://fb.workplace.com/groups/hackforhiphop/permalink/24988113060810640/) or D47890092). This diff continues the implementation of a new strict switch gated behind an annotation.

It does so by adding refactoring the logic for deciding whether or not a newtype is an enum. Depending on whether it is an enum, enum class, or enum class label, it now returns all the relevant information in the variant itself, rather than as arguments to a callback. This will simplifiy an upcoming diff for adding support for enum classes, and enum class labels.

Reviewed By: dlreeves

Differential Revision: D49775754

fbshipit-source-id: 219bf9fcf02fa3ecd029c819d9e8f0f4ee25f89d