Add support for enums
Summary:
Exhaustivity checking in Hack misses many cases (see [this post](https://fb.workplace.com/groups/hackforhiphop/permalink/24988113060810640/) or D47890092). This diff continues the implementation of a new strict switch gated behind an annotation.

It does so by adding support for enums (plain ones, not enum classes or enum class labels). To achieve this, two constructors have been added: `AllEnums` and `Enum of EnumInfo.t`. The first is used when generating a set of values for `Tarraykey`; the second is used for any actual enum. To handle the fact that symbolically, `Enum info <: AllEnums`, `Value.inter` has been added, which performs an _approximate_ greatest-lower-bound/meet.

Note intersections such as `Enum1 & Enum2`, `Enum & string` (or `int & Enum`) are simplified to `{}`. These will be dealt with in an upcoming diff.

Enum constants are mapped to multiple keys `[Enum info; AllEnum]`. This means that when partitioning cases based on type, an enum constant will be attempted to be bucketed into either or both `Enum info` and `AllEnum`. Only if neither bucket exists, will the constant be counted as the wrong type. This is the reason for the `List.exists` in `partition_cases`.

Reviewed By: dlreeves

Differential Revision: D49682779

fbshipit-source-id: 0cfcb253c5d18bc53f9e339c6f7ffb58dc40f7bb