Optimized member fanout: ignore private changes
Summary:
The current production algorithm ignores the fanout of private member changes, and in this diff we do the same for the new algorithm.

Note 1: in reality, the production algorithm only considers the change "private" if the member is added or removed, but not if it is modified. Odd fluke. We plan to fix that in a future diff, but gated by a different flag.

Note 2: The reason I had not done that before is because I thought you could still override private members with public members. Turns out you sort of can but it does not really override it, as shown by new test test/typecheck/private_override.php. So ignoring the fanout of private members is correct.

Reviewed By: rjbailey

Differential Revision: D49911763

fbshipit-source-id: f884cc33f305cc7a3458fb1a08b68314f07b001e