Add support for dynamic type
Summary:
Exhaustivity checking in Hack misses many cases (see [this post](https://fb.workplace.com/groups/hackforhiphop/permalink/24988113060810640/) or D47890092). This diff continues the implementation of a new strict switch gated behind an annotation.

It does so by adding support for `dynamic` types. This is done in a way that `dynamic | T` i.e. `~T` is always treated as `T`, except for when `T = noreturn` (i.e. just `dynamic`). In the latter case, a default is required and all other cases are rejected.

Reviewed By: dlreeves

Differential Revision: D48955062

fbshipit-source-id: 41f3bc00b276e3a4245417d774e3e97137628352