Store whether a unit is soft deployed in repo mode
Summary:
In repo mode, a unit's **package accessibility** can be one of three options:
- Hard deployed: included in the current deployment, accessible
- Soft deployed: included in the current deployment, warns when accessed
- Not deployed: not included in current deployment, raises undefined

Note that in the first case, we don't need to check package deployment,  and in the third case in repo mode, we'll throw an undefined error on loading of the symbol (and thus, won't check the package violation at all). Therefore, in repo mode, the only case that requires us to check a unit's package is if the unit is soft deployed.

This check is currently very slow to check when we cannot statically deduce a symbol's caller (i.e. in dynamic callsites). So to optimize this, in repo mode, we store a single boolean in the unit that is true if and only if the unit is soft deployed. We can then check that boolean in the dynamic case.

Implementation notes:
- I'd love to add a single bitfield instead of a full boolean to each unit, but it's undefined behavior in C++ to get the offset of a bitfield, and we need to be able to get the offset to access it without doing a callq. If there's another way to do this (that doesn't involve, say, hackily calculating the offset of the bitfield with some garbage like `reinterpret_cast<std::size_t>(&nullPtr->m_softDeployedRepoOnly)`), let me know.

- The field is named as m_softDeployedRepoOnly to make sure it is only accessed in repoAuthoritative mode, and its getter asserts that we're in repo mode only. However, the offset of m_softDeployedRepoOnly is a static constexpr, so I don't really have a way of asserting that the offset is only accessed in repo mode. Would appreciate feedback.

Reviewed By: oulgen

Differential Revision: D49873700

fbshipit-source-id: caa683fb8410fb156f8040a65b5458249c8a5078