add "convert to 'var_dump'" quickfix
Summary:
Change: add a new quickfix for converting `=$x + 1` to valid Hack: `var_dump($x + 1)`

Mitigates  T164253172 when T165204669 is fixed

## Why

- statements with leading `=` like `=$x + 1;` are only allowed in the debugger.
- When users paste from the debugger to Code@FB or Bento notebooks, they get syntax errors
- This change reduces the pain of `=$x + 1` not being allowed

## Perf note

This change introduces some repeat work in parsing but *only* in the case where we have an error about an unexpected `=`, which I doubt is common.

## Advice Welcome

See leading_equals_expr_1.php: we offer a quickfix for changing `foo(=$x)` to `foo(var_dump($x))` and I have a weak preference for not doing so. I think the quickfix is unlikely to be what users want when `=` starts an expression.

Is there a way to not provide the refactoring for cases like `foo(=$x)` without threading extra information through a lot of function calls in the expression parser (`is_top_of_statement` or some such)?

## Follow-on work

**doesn't actually show up in IDE**

There was an off-by-one in our calculation (in our language server) of which quickfixes to show to the user. Fixed in {D49772308}

**no quickfixes at all show up in notebooks**

T165204669

**Too many flavors of Hack syntax**

We currently have at least three flavors of Hack syntax:
- actual Hack: no syntax errors in IDE nor HHVM
- debugger: unspecified
- notebooks (see hh_rewrite_declarations for part of the definition): unspecified

Ideally we will conform and specify where feasible to increase maintainability and learnability-for-aend-users. Looking into it.

See https://www.internalfb.com/diff/D49687153?dst_version_fbid=1768488200255193&transaction_fbid=6924815540915718 for why we seriously consider allowing leading `=` in notebooks

Reviewed By: patriciamckenzie

Differential Revision: D49687153

fbshipit-source-id: 1b742af585658f0227016fc1f89cb02b20fe6039