HH\class_to_classname and classname_to_class helpers
Summary: Adds two `HH\` functions for converting between lazy classes / class pointers and strings at runtime. For class pointer creation, we raise a notice if the class isn't annotated with the `__DynamicallyReferenced` attribute -- sampled until we land on the exact semantics we want e.g. is `__DR` inherited or not.

Reviewed By: arnabde03

Differential Revision: D48288075

fbshipit-source-id: 6230944de67b55243cabe2f7789e1f1910561fa1