rename extract method no experimental capabilities tests
Summary:
## Description

No user-facing change.

Rename some tests because we add non-extract-method tests for snippet removal in {D49594353}.

## Context:

We have some tests for what happens when a Language Server Protocol client does not support experimental snippet text edits.

Snippet text edits are what allow us to control placeholders for text inserted by code actions. See the video in the summary for D49581418 where we send the cursor position to the middle of the doc comment.

The tests here and in the next diff ({D49594353}) ensure that we do not include snippets in code actions when the client doesn't support snippets in code actions. Non-Code@FB clients almost certainly will not support our nonstandard extension to LSP

Differential Revision: D49594354

fbshipit-source-id: 67873dbbcb2c40105d32c2aa847aab4ece5616e9