test for remove snippet
Summary:
## Description

Add a test illustrating a bug: we fail to remove snippet text edits for the new "Add doc comment" refactor.

The bug is fixed in the next diff in the stack: ({D49594351})

## Context (same as previous diff):

We have some tests for what happens when a Language Server Protocol client does not support experimental snippet text edits.

Snippet text edits are what allow us to control placeholders for text inserted by code actions. See the video in the summary for D49581418 where we send the cursor position to the middle of the doc comment.

The tests here and in the next diff ({D49594353}) ensure that we do not include snippets in code actions when the client doesn't support snippets in code actions. Non-Code@FB clients almost certainly will not support our nonstandard extension to LSP

Reviewed By: patriciamckenzie

Differential Revision: D49594353

fbshipit-source-id: ba07c751363235d88df80fea898f9107ea9accbd