Specific parse error for statement that starts with `=`
Summary:
Make a specific parse error for statement that starts with `=`, rather than the general parse error we had before.

Context:  the following is allowed in the debugger but is not valid Hack:

```
x = 2;
=$x;
```

By providing a specific error code and message, I expect we will make understanding the error easier for users of IDE and `hh` because they will see a more meaningful error message when pasting from the debugger to Code@FB/Vim+LSP/Bento/etc.

Additional reasoning in the fb-only section

Differential Revision: D49914614

fbshipit-source-id: 720e22ea456844e06a6f239765506b2e144ef1fd